refactor: move kernel plugins to separate files
